+++ src/fluctus.cpp
@@ -47,6 +47,7 @@ struct Fluctus : SanguineModule {
 		PARAM_HI_FI,
 		PARAM_STEREO,
 		PARAM_LEDS_MODE,
+		PARAM_OUT_GAIN,
 		PARAMS_COUNT
 	};
 	enum InputIds {
@@ -187,6 +188,8 @@ struct Fluctus : SanguineModule {
 		configParam(PARAM_REVERB, 0.f, 1.f, 0.5f, "Reverb", "%", 0.f, 100.f);
 
 		configParam(PARAM_IN_GAIN, 0.f, 1.f, 0.5f, "Input gain", "%", 0.f, 100.f);
+
+		configParam(PARAM_OUT_GAIN, 0.f, 2.f, 1.f, "Output gain", "%", 0.f, 100.f);
 
 		configInput(INPUT_LEFT, "Left");
 		configInput(INPUT_RIGHT, "Right");
@@ -361,9 +364,11 @@ struct Fluctus : SanguineModule {
 		if (!outputBuffer.empty()) {
 			outputFrame = outputBuffer.shift();
 			if (outputs[OUTPUT_LEFT].isConnected()) {
+				outputFrame.samples[0] *= params[PARAM_OUT_GAIN].getValue();
 				outputs[OUTPUT_LEFT].setVoltage(5.f * outputFrame.samples[0]);
 			}
 			if (outputs[OUTPUT_RIGHT].isConnected()) {
+				outputFrame.samples[1] *= params[PARAM_OUT_GAIN].getValue();
 				outputs[OUTPUT_RIGHT].setVoltage(5.f * outputFrame.samples[1]);
 			}
 		}
@@ -720,11 +725,13 @@ struct FluctusWidget : SanguineModuleWidget {
 
 		addParam(createParamCentered<Sanguine1PYellow>(millimetersToPixelsVec(125.214, 96.727), module, Fluctus::PARAM_REVERB));
 
-		addParam(createParamCentered<Rogan1PWhite>(millimetersToPixelsVec(14.603, 97.272), module, Fluctus::PARAM_IN_GAIN));
+		addParam(createParamCentered<Rogan1PWhite>(millimetersToPixelsVec(41.434, 117.002), module, Fluctus::PARAM_IN_GAIN));
 
-		addParam(createLightParamCentered<VCVLightLatch<MediumSimpleLight<GreenLight>>>(millimetersToPixelsVec(31.438, 97.272),
+		addParam(createParamCentered<Rogan1PWhite>(millimetersToPixelsVec(95.701, 117.002), module, Fluctus::PARAM_OUT_GAIN));
+
+		addParam(createLightParamCentered<VCVLightLatch<MediumSimpleLight<GreenLight>>>(millimetersToPixelsVec(18.631, 96.727),
 			module, Fluctus::PARAM_HI_FI, Fluctus::LIGHT_HI_FI));
-		addParam(createLightParamCentered<VCVLightLatch<MediumSimpleLight<RedLight>>>(millimetersToPixelsVec(56.63, 97.272),
+		addParam(createLightParamCentered<VCVLightLatch<MediumSimpleLight<RedLight>>>(millimetersToPixelsVec(43.823, 96.727),
 			module, Fluctus::PARAM_STEREO, Fluctus::LIGHT_STEREO));
 
 		addInput(createInputCentered<BananutGreen>(millimetersToPixelsVec(7.677, 116.972), module, Fluctus::INPUT_LEFT));