+++ src/nebulae.cpp
@@ -54,6 +54,7 @@ struct Nebulae : SanguineModule {
 		PARAM_HI_FI,
 		PARAM_STEREO,
 		PARAM_LEDS_MODE,
+		PARAM_OUT_GAIN,
 		PARAMS_COUNT
 	};
 	enum InputIds {
@@ -190,6 +191,8 @@ struct Nebulae : SanguineModule {
 		configParam(PARAM_REVERB, 0.f, 1.f, 0.5f, "Reverb", "%", 0.f, 100.f);
 
 		configParam(PARAM_IN_GAIN, 0.f, 1.f, 0.5f, "Input gain", "%", 0.f, 100.f);
+
+		configParam(PARAM_OUT_GAIN, 0.f, 2.f, 1.f, "Output gain", "%", 0.f, 100.f);
 
 		configInput(INPUT_LEFT, "Left");
 		configInput(INPUT_RIGHT, "Right");
@@ -359,9 +362,11 @@ struct Nebulae : SanguineModule {
 		if (!outputBuffer.empty()) {
 			outputFrame = outputBuffer.shift();
 			if (outputs[OUTPUT_LEFT].isConnected()) {
+				outputFrame.samples[0] *= params[PARAM_OUT_GAIN].getValue();
 				outputs[OUTPUT_LEFT].setVoltage(5.f * outputFrame.samples[0]);
 			}
 			if (outputs[OUTPUT_RIGHT].isConnected()) {
+				outputFrame.samples[1] *= params[PARAM_OUT_GAIN].getValue();
 				outputs[OUTPUT_RIGHT].setVoltage(5.f * outputFrame.samples[1]);
 			}
 		}
@@ -681,11 +686,13 @@ struct NebulaeWidget : SanguineModuleWidget {
 
 		addParam(createParamCentered<Sanguine1PYellow>(millimetersToPixelsVec(125.214, 96.727), module, Nebulae::PARAM_REVERB));
 
-		addParam(createParamCentered<Rogan1PWhite>(millimetersToPixelsVec(14.603, 97.272), module, Nebulae::PARAM_IN_GAIN));
+		addParam(createParamCentered<Rogan1PWhite>(millimetersToPixelsVec(41.434, 117.002), module, Nebulae::PARAM_IN_GAIN));
 
-		addParam(createLightParamCentered<VCVLightLatch<MediumSimpleLight<GreenLight>>>(millimetersToPixelsVec(31.438, 97.272),
+		addParam(createParamCentered<Rogan1PWhite>(millimetersToPixelsVec(95.701, 117.002), module, Nebulae::PARAM_OUT_GAIN));
+
+		addParam(createLightParamCentered<VCVLightLatch<MediumSimpleLight<GreenLight>>>(millimetersToPixelsVec(18.631, 96.727),
 			module, Nebulae::PARAM_HI_FI, Nebulae::LIGHT_HI_FI));
-		addParam(createLightParamCentered<VCVLightLatch<MediumSimpleLight<RedLight>>>(millimetersToPixelsVec(56.63, 97.272),
+		addParam(createLightParamCentered<VCVLightLatch<MediumSimpleLight<RedLight>>>(millimetersToPixelsVec(43.823, 96.727),
 			module, Nebulae::PARAM_STEREO, Nebulae::LIGHT_STEREO));
 
 		addInput(createInputCentered<BananutGreen>(millimetersToPixelsVec(7.677, 116.972), module, Nebulae::INPUT_LEFT));